fix: make container deletion more robust
Use the `IsErrNotFound` function exported from the Docker client library
to detect HTTP 404 errors returned by the Docker API more reliably.

This improves compatibility with Podman's implementation of the Docker
API, because Podman uses a different capitalization for the error
message that is returned when a container has already been deleted,
causing the existing code to fail to recognize this condition.